+++ CHANGES.md
@@ -1,6 +1,7 @@
 # 7.2.1 (2024-11-14)
 
 * add official support for floating point values in ColorMap
+* cast data to `uint8` datatype when applying linear colormap
 
 # 7.2.0 (2024-11-05)